Async ImageLoader (#304)
* add async loader

* using shuffle channel

+++ example/dcgan/dcgan.go
@@ -93,7 +93,7 @@ func discriminator(nc int64, ndf int64) *nn.SequentialModule {
 	)
 }
 
-func celebaLoader(data string, vocab map[string]int64, mbSize int) *datasets.ImageLoader {
+func celebaLoader(data string, vocab map[string]int, mbSize int) *datasets.ImageLoader {
 	imageSize := 64
 	trans := transforms.Compose(transforms.Resize(imageSize),
 		transforms.CenterCrop(imageSize),
